feat: 타입 정리
