+++ apps/nowait-user/src/pages/waiting/bookmark/components/BookmarkEmptyPage.tsx
@@ -1,13 +1,17 @@
-import { SmallActionButton } from '../../../../components/SmallActionButton'
-import { useNavigate } from 'react-router-dom'
-import Add from "../../../../assets/icon/Add.svg?react"
+import { SmallActionButton } from "../../../../components/SmallActionButton";
+import { useNavigate } from "react-router-dom";
+import Add from "../../../../assets/icon/Add.svg?react";
 
 const BookmarkEmptyPage = () => {
-    const navigate = useNavigate()
+  
+  const navigate = useNavigate();
+
   return (
     <div className="min-h-screen flex flex-col justify-center items-center bg-white">
       <h1 className="whitespace-pre-line text-16-regular text-black-80 mb-5 text-center">
-        아직 북마크한 부스가 없어요.<br/>마음에 드는 부스를 찾아보세요!
+        아직 북마크한 부스가 없어요.
+        <br />
+        마음에 드는 부스를 찾아보세요!
       </h1>
       <SmallActionButton
         mode="default"
@@ -19,7 +23,7 @@ const BookmarkEmptyPage = () => {
         <Add className="w-4 h-4" fill="currentColor" />
       </SmallActionButton>
     </div>
-  )
-}
+  );
+};
 
-export default BookmarkEmptyPage
+export default BookmarkEmptyPage;
