+++ apps/nowait-user/src/types/wait/bookmark.ts
@@ -1,0 +1,9 @@
+// 북마크 생성 시 응답 데이터 타입
+export interface BookmarkResponse {
+  success: boolean;
+  response: {
+    bookmarkId: string;
+    userId: string;
+    storeId: string;
+  };
+}
